refactor ethercat code
