+++ march_ethercat/ethercat_SDO.h
@@ -1,0 +1,11 @@
+#ifndef SDO_
+#define SDO_
+
+#include <math.h>
+#include <stdint.h>
+
+int sdo_bit8(int slave, uint32_t index, uint8_t sub, uint8_t value);
+int sdo_bit16(int slave, uint32_t index, uint8_t sub, uint16_t value);
+int sdo_bit32(int slave, uint32_t index, uint8_t sub, uint32_t value);
+
+#endif
