--- march_ethercat/ros_init.cpp
@@ -1,198 +1,0 @@
-#include "ros_init.h"
-
-extern "C" {
-#include "osal.h"
-}
-
-//------------------------------------------------------------------------------------------------//
-// JOINT HANDLER PUB FUNCTIONS
-//------------------------------------------------------------------------------------------------//
-
-void publish_to_joint_handler(string slaveName, custom_msgs::ECtoJH msg)
-{
-  (jointHandlerPubMap.find(slaveName)->second)->publish(msg);
-}
-
-//------------------------------------------------------------------------------------------------//
-// GES HANDLER PUB FUNCTIONS
-//------------------------------------------------------------------------------------------------//
-
-void publish_to_ges_handler(string slaveName, custom_msgs::ECtoUG msg)
-{
-  (gesHandlerPubMap.find(slaveName)->second)->publish(msg);
-}
-
-void publish_to_ges_handler(string slaveName, custom_msgs::ECtoLG msg)
-{
-  (gesHandlerPubMap.find(slaveName)->second)->publish(msg);
-}
-
-// void publish_to_ges_handler(string slaveName, custom_msgs::ECtoBPG msg)
-// {
-// 	(gesHandlerPubMap.find(slaveName)->second)->publish(msg);
-// }
-
-//------------------------------------------------------------------------------------------------//
-// GES HANDLER PUB FUNCTIONS
-//------------------------------------------------------------------------------------------------//
-
-void publish_to_safety(custom_msgs::data8Msg msg)
-{
-  safetyPubPtr->publish(msg);
-}
-
-//------------------------------------------------------------------------------------------------//
-// GES HANDLER PUB FUNCTIONS
-//------------------------------------------------------------------------------------------------//
-
-void publish_to_pdb_handler(custom_msgs::ECtoPDB msg)
-{
-  PdbHandlerPubPtr->publish(msg);
-}
-
-void publish_to_ipd_handler(custom_msgs::ECtoIPD msg)
-{
-  IpdHandlerPubPtr->publish(msg);
-}
-
-// ros spinOnce call
-void rosSpinOnce()
-{
-  // pulls from subscribed topics and executes callback functions
-  ros::spinOnce();
-}
-
-void template_ges_value_callback(std_msgs::UInt8 msg)
-{
-    printf("Set data to %i\n",msg.data);
-    Template_ges_callback::set_ges_data("TEMPL_GES",msg.data);
-}
-
-void rosloop(ros::NodeHandle nh)
-{
-  //------------------------------------------------------------------------------------------------//
-  // JOINT HANDLER PUBS & SUBS
-  //--//--------------------------------------------------------------------------------------------//
-  // 	UPSTREAM COMBINED MESSAGES
-  //--------------------------------------------------------------------------------------------//
-
-  for (int i = 0; i < *LaunchParameters::get_number_of_joints(); i++)
-  {
-    // create publishers w/ datatype, topic name and queue size
-    string prefix = (*LaunchParameters::get_list_of_joints())[i];
-
-    jointHandlerPubMap[prefix] = (new ros::Publisher(nh.advertise<custom_msgs::ECtoJH>("ECto" + prefix, 50)));
-
-    //--------------------------------------------------------------------------------------------//
-    // TARGET POSITION
-    //--------------------------------------------------------------------------------------------//
-    jointPositionSubList.push_back(
-        new ros::Subscriber(nh.subscribe(prefix + "TargetPositionIU", 1, &Imc_callback::set_target_position)));
-
-    //--------------------------------------------------------------------------------------------//
-    // CONTROLWORD
-    //--------------------------------------------------------------------------------------------//
-    jointWordSubList.push_back(
-        new ros::Subscriber(nh.subscribe(prefix + "Controlword", 50, &Imc_callback::set_controlword)));
-  };
-
-  //------------------------------------------------------------------------------------------------//
-  // GES HANDLER PUBS & SUBS
-  //--//--------------------------------------------------------------------------------------------//
-  // COMBINED MESSAGES UPSTREAM
-  //--------------------------------------------------------------------------------------------//
-
-  for (vector<string>::iterator i = LaunchParameters::get_list_of_GES()->begin();
-       i != LaunchParameters::get_list_of_GES()->end(); ++i)
-  {
-    if (*i == "LUG" || *i == "RUG")
-    {
-      gesHandlerPubMap[*i] = (new ros::Publisher(nh.advertise<custom_msgs::ECtoUG>("ECto" + *i, 50)));
-    }
-    else if (*i == "LLG" || *i == "RLG")
-    {
-      gesHandlerPubMap[*i] = (new ros::Publisher(nh.advertise<custom_msgs::ECtoLG>("ECto" + *i, 50)));
-    }
-    else if (*i == "BPG")
-    {
-      // gesHandlerPubMap[*i] = (new ros::Publisher(nh.advertise<custom_msgs::ECtoBPG>("ECto" + *i, 50)));
-
-      speakerSubPtr = (new ros::Subscriber(nh.subscribe(*i + "toEC", 50, &Backpack_ges_callback::send_frequency)));
-    }
-    else if (*i == "TEMPL_GES")
-    {
-      // Todo: Implement listening to led command topic
-      printf("Set callback for listening to led command topic\n");
-    }
-    else
-    {
-      printf("error when making ges publishers! unknown name\n");
-    }
-  }
-
-  //------------------------------------------------------------------------------------------------//
-  // SAFETY PUBS & SUBS
-  //--//--------------------------------------------------------------------------------------------//
-  // SLAVE LOST ERROR
-  //--------------------------------------------------------------------------------------------//
-
-  safetyPubPtr = (new ros::Publisher(nh.advertise<custom_msgs::data8Msg>("ECErrorMessage", 50)));
-
-  //------------------------------------------------------------------------------------------------//
-  // PDB HANDLER PUBS & SUBS
-  //--//--------------------------------------------------------------------------------------------//
-  // UPSTREAM COMBINED MESSAGES
-  //--------------------------------------------------------------------------------------------//
-  if (*LaunchParameters::get_number_of_Pdb())
-  {
-    PdbHandlerPubPtr = (new ros::Publisher(nh.advertise<custom_msgs::ECtoPDB>("ECtoPDB", 50)));
-
-    //--------------------------------------------------------------------------------------------//
-    // PDB command
-    //--------------------------------------------------------------------------------------------//
-
-    PdbCommandSubPtr = (new ros::Subscriber(nh.subscribe("PDBtoEC", 50, &Pdb_callback::set_pdb_command)));
-  }
-
-  //------------------------------------------------------------------------------------------------//
-  // IPD HANDLER PUBS & SUBS
-  //--//--------------------------------------------------------------------------------------------//
-  // UPSTREAM COMBINED MESSAGES
-  //--------------------------------------------------------------------------------------------//
-
-  if (*LaunchParameters::get_number_of_ipd())
-  {
-    IpdHandlerPubPtr = (new ros::Publisher(nh.advertise<custom_msgs::ECtoIPD>("ECtoIPD", 50)));
-
-    //--------------------------------------------------------------------------------------------//
-    // IPD DOWN
-    //--------------------------------------------------------------------------------------------//
-
-    IpdHandlerSubPtr = (new ros::Subscriber(nh.subscribe("IPDtoEC", 50, &Ipd_callback::set_ipd_data)));
-  }
-
-  if (*LaunchParameters::get_number_of_GES())
-  {
-    TemplateHandlerSubPtr = (new ros::Subscriber(nh.subscribe("march/template/data", 50, &template_ges_value_callback)));
-  }
-
-  cout << "number of joints: " << *LaunchParameters::get_number_of_joints() << endl;
-  cout << "number of ges: " << *LaunchParameters::get_number_of_GES() << endl;
-  cout << "number of pdb: " << *LaunchParameters::get_number_of_Pdb() << endl;
-  cout << "number of ipd: " << *LaunchParameters::get_number_of_ipd() << endl;
-
-  ros::Rate rate(*LaunchParameters::get_ethercat_frequency());
-  while (ros::ok())
-  {
-    // run a SOEM loop and publish data
-    update();
-
-    // receive and act on subscribed nodes
-    ros::spinOnce();
-
-    // wait unitl next iteration unless over time
-    rate.sleep();
-  };
-
-  cout << "escape!" << endl;
-}
