+++ march_ethercat/soem/osal.h
@@ -44,7 +44,7 @@ boolean osal_timer_is_expired(osal_timert* self);
 int osal_usleep(uint32 usec);
 ec_timet osal_current_time(void);
 void osal_time_diff(ec_timet* start, ec_timet* end, ec_timet* diff);
-int march_osal_thread_create(void* thandle, int stacksize, void (*func)(void*), char* (*funcPtr)(const long int*));
+//int march_osal_thread_create(void* thandle, int stacksize, void (*func)(void*), char* (*funcPtr)(const long int*));
 int osal_thread_create(void* thandle, int stacksize, void* func, void* param);
 int osal_thread_create_rt(void* thandle, int stacksize, void* func, void* param);
